Moved dataflow tests into a module
